requirements.txt
Pillow==10.0.0
rembg==2.0.50
python-telegram-bot==20.4